feat: add makefile
