+++ k8s/log/log
@@ -1,6 +1,9 @@
 #!/bin/bash
 
-KUBE_LOGGER_SCRIPT="$SERVICE_PATH/log/kube-logger-bash/exec"
+PLATFORM=$(uname | tr '[:upper:]' '[:lower:]')
+ARCH=$(uname -m)
+
+KUBE_LOGGER_SCRIPT="$SERVICE_PATH/log/kube-logger-go/bin/$PLATFORM/log-$ARCH"
 
 if [ ! -f "$KUBE_LOGGER_SCRIPT" ]; then
     echo "Error: kube-logger bash script not found at $KUBE_LOGGER_SCRIPT" >&2
